requirements.txt
streamlit>=1.28.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.3
plotly>=5.15.0
PyPDF2>=3.0.1
python-docx>=0.8.11